Refactor: content pins into luria.pins with an extensible source table
The pin machinery moves out of remotes.py into its own module, and the
'where do stable bytes live' question becomes SOURCES — an ordered table
of rungs (declared pin_url template, then the GitHub raw rebase), each
answering a URL or ''. A new source-specific case is one function and
one table entry; everything downstream consumes only the returned URL.
Tests split into tests/test_pins.py, sharing test_remotes' fixtures.
